requirements.txt
fastapi 
uvicorn
openai-whisper 
ffmpeg-python 
srt
jinja2 
python-multipart 
aiofiles
yt-dlp 
ffmpeg